# docs/html/spec-files/swagger.yaml
swagger: '2.0'
info:
  version: 1.0.0
  title: 加盟商API
  description: |

    __所有时间戳为微秒(ms)__


    > 所有请求结果都包含如下字段(接口定义中不再声明)

    ```
    {
        success: boolean,
        errorCode: integer,
        errorMessage: string,
        data,
    }
    ```

    >当`success`为`true`时, `errorCode`和`errorMessage`为空.
    >当`success`为`false`间, `data`为空
basePath: /api
schemes:
  - http
produces:
  - application/json
consumes:
  - application/json
...
paths:
  /auth/login:
    post:
      operationId: 登陆
      parameters:
        - in: body
          name: body
          schema:
            type: object
            properties:
              phone:
                type: string
              password:
                type: string
          required: true
      responses:
        '200':
          schema:
            type: object
            properties:
              data:
                $ref: '#/definitions/LoginCredential'
          description: 请求成功
      summary: 登陆
...
  /login:
    post:
      operationId: 登陆
      parameters:
        - in: body
          name: body
          schema:
            type: object
            properties:
              username:
                type: string
                j-required: true
              password:
                type: string
                j-required: true
          required: true
      responses:
        '200':
          schema:
            type: object
            properties:
              data:
                type: string
                m-type: hex
                m-length: 32
          description: 请求成功
      summary: 登陆
...
definitions:
...
  LoginCredential:
    type: object
    properties:
      name:
        type: string
...